Closes OOZIE-93 coordinator start and end instances doesn't work with float number.
From: Mohammad Kamrul Islam <kamrul@theybeside-lm.(none)>

git-svn-id: https://svn.apache.org/repos/asf/incubator/oozie/trunk@1168869 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/oozie/command/coord/CoordCommandUtils.java
@@ -75,7 +75,7 @@ public class CoordCommandUtils {
         if (firstPos >= 0 && lastPos > firstPos) {
             String tmp = funcName.substring(firstPos + 1, lastPos).trim();
             if (tmp.length() > 0) {
-                return Integer.parseInt(tmp);
+                return (int) Double.parseDouble(tmp);
             }
         }
         throw new RuntimeException("Unformatted function :" + funcName);
@@ -89,7 +89,7 @@ public class CoordCommandUtils {
             String tmp = funcName.substring(firstPos + 1, secondPos).trim();
             if (tmp.length() > 0) {
                 restArg.append(funcName.substring(secondPos + 1, lastPos).trim());
-                return Integer.parseInt(tmp);
+                return (int) Double.parseDouble(tmp);
             }
         }
         throw new RuntimeException("Unformatted function :" + funcName);
